Optimize CRYPT computations (issue #1)
Since password checks are incredibly expensive, they should always be paired (AND) with a condition that selects a very small number of records (ideally just one).

However, the Postgres query analyzer isn't capable of properly optimizing the checks so that the crypt calculations are done at the end (on the smallest number of records).

This changes the grammar to enforce users to always AND a password check with another (complex) condition, as well as changing the SQL generator to produce a query that performs the CRYPT checks on the result of the AND term.