Add EXPOSE_SCIM_EXCEPTIONS setting (#64)
* Add EXPOSE_SCIM_EXCEPTIONS setting

This commit adds an EXPOSE_SCIM_EXCEPTIONS setting that allows users of
the library to toggle whether SCIM errors are shown to the scim client.

In some circumstances it can be beneficial for the client
to know what caused an error. However, this can present an
unacceptable security risk for many companies. This flag
allows for a generic error message to be returned when such a
security risk is unacceptable. The flag defaults to False, thus not
displaying errors to clients.

* Update docs to include EXPOSE_SCIM_EXCEPTIONS

This commit updates the docs to include all settings and an explanation
for the EXPOSE_SCIM_EXCEPTIONS setting.