Move completely to Poetry
This commit removes the setup.py file and updates the
pyproject.yaml and tox.ini files to support a completely Poetry based
workflow. This change was made to better support the rapid development
of fixes amidst the ever changing developer tool chain landscape.

This commit also drops support for Django <= 3.2. Security support
ended on 2022-04-01.

This commit adds testing support for Python 3.10.

This commit disables the calculation of Coverage. Coverage will be
reintroduced after migration to pytest. Coverage is not enabled now
(with Poetry alone) given the difficulty and time required to do so.
That time is better spent on migrating to pytest where this coverage
calculation would be trivial.